datapath: Move DeviceManager to the datapath package
datapath/linux package is a more natural location for the
device detection code now that runtime detection will
introduce dependencies to DeviceManager from other packages.
